本番環境での修正

+++ db/migrate/20241216005109_add_group_id_to_group_tags.rb
@@ -1,6 +1,6 @@
 class AddGroupIdToGroupTags < ActiveRecord::Migration[6.1]
   def change
-    add_column :group_tags, :group_id, :integer
-    add_index :group_tags, :group_id # インデックスの追加（必要に応じて）
+    add_column :group_tags, :group_id, :bigint  # 修正：bigint型に変更
+    add_index :group_tags, :group_id  # インデックスの追加（必要に応じて）
   end
 end
