+++ db/migrate/20241220012949_add_prefecture_id_to_users.rb
@@ -1,6 +1,6 @@
 class AddPrefectureIdToUsers < ActiveRecord::Migration[6.1]
   def change
-    add_column :users, :prefecture_id, :integer, null: false
+    add_column :users, :prefecture_id, :bigint, null: false
     add_foreign_key :users, :prefectures
   end
 end
